requirements.txt
flask==3.0.0
flask-socketio==5.3.6
requests==2.31.0
python-dotenv==1.0.0
pyttsx3==2.90
speechrecognition==3.10.0
pyaudio==0.2.11
pytesseract==0.3.10
Pillow==10.1.0
psutil==5.9.6
cryptography==41.0.7
PyQt5==5.15.10
PyQtWebEngine==5.15.6
python-docx==0.8.11
docx2pdf==0.3.3
plyer==2.1.0
pyperclip==1.8.2
librosa==0.10.1
numpy==1.24.3
apscheduler==3.10.4
porcupine-api==2.1.0  # Requires Picovoice account
gunicorn==21.2.0